+++ src/views/AccountDetails/AccountDetails.js
@@ -8,25 +8,26 @@ import {
 } from 'react-intl';
 
 import {
-  Paneset,
-  Pane,
   Button,
-  Row,
   Col,
   KeyValue,
   MultiColumnList,
   NoValue,
+  Pane,
+  Paneset,
+  Row,
 } from '@folio/stripes/components';
 
 import Actions from '../../components/Accounts/Actions/FeeFineActions';
 import {
-  getFullName,
   calculateSortParams,
+  getFullName,
 } from '../../components/util';
 
 import {
   calculateTotalPaymentAmount,
   isRefundAllowed,
+  isCancelAllowed,
 } from '../../components/Accounts/accountFunctions';
 
 import css from './AccountDetails.css';
@@ -278,6 +279,7 @@ class AccountDetails extends React.Component {
 
     const isAccountsPending = _.get(resources, ['accounts', 'isPending'], true);
     const isActionsPending = _.get(resources, ['accountActions', 'isPending'], true);
+    const feeFineActions = _.get(resources, ['accountActions', 'records'], []);
 
     const actions = this.state.data || [];
     const actionsSort = _.orderBy(actions, [this.sortMap[sortOrder[0]], this.sortMap[sortOrder[1]]], sortDirection);
@@ -293,8 +295,9 @@ class AccountDetails extends React.Component {
     const lostItemPolicyName = itemDetails?.lostItemPolicyName;
     const contributors = itemDetails?.contributors.join(', ');
 
-    const totalPaidAmount = calculateTotalPaymentAmount(resources?.accounts?.records);
-    const refundAllowed = isRefundAllowed(account);
+    const totalPaidAmount = calculateTotalPaymentAmount(resources?.accounts?.records, feeFineActions);
+    const refundAllowed = isRefundAllowed(account, feeFineActions);
+    const cancelAllowed = isCancelAllowed(account);
 
     return (
       <Paneset isRoot>
@@ -346,7 +349,7 @@ class AccountDetails extends React.Component {
               </Button>
               <Button
                 id="errorAccountActionsHistory"
-                disabled={disabled || buttonDisabled || isActionsPending || isAccountsPending}
+                disabled={disabled || buttonDisabled || isActionsPending || isAccountsPending || !cancelAllowed}
                 buttonStyle="primary"
                 onClick={this.error}
               >