+++ test/bigtest/network/config.js
@@ -147,6 +147,8 @@ export default function config() {
       const filterField = 'active';
       const departmentsField = 'departments';
       const usernameField = 'username';
+      const lastNameField = 'personal.lastName';
+      const barcodeField = 'barcode';
       if (/^%28/.test(query)) {
         query = decodeURIComponent(query);
       }
@@ -172,6 +174,16 @@ export default function config() {
       if (field === usernameField) {
         return users.where({
           [usernameField]: term.replace('*', '')
+        });
+      }
+      if (field === lastNameField) {
+        return users.where(u => {
+          return (u.personal.lastName === term.replace('*', ''));
+        });
+      }
+      if (field === barcodeField) {
+        return users.where({
+          [barcodeField]: term.replace('*', '')
         });
       }
     }
@@ -433,6 +445,7 @@ export default function config() {
   this.post('/perms/users/:id/permissions?indexField=userId');
 
   this.post('/circulation/loans/:loanId/declare-item-lost', []);
+  this.post('/circulation/loans/:loanId/claim-item-returned', []);
 
   this.get('/feefineactions', ({ feefineactions }) => {
     return this.serializerOrRegistry.serialize(feefineactions.all());
@@ -855,6 +868,61 @@ export default function config() {
         },
       ],
       amount : 40.00
+    };
+  });
+
+  this.post('/accounts-bulk/check-refund', ({ accounts }, request) => {
+    return {
+      amounts: [
+        {
+          accountId: 'r1',
+          amount: 100
+        },
+        {
+          accountId: 'r2',
+          amount: 200
+        },
+        {
+          accountId: 'r3',
+          amount: 100
+        }
+      ],
+      allowed: true,
+      amount: 400,
+      remainingAmount: 100
+    };
+  });
+
+  this.post('/accounts-bulk/refund', ({ accounts }, request) => {
+    return {
+      accountIds: ['r1', 'r2', 'r3'],
+      feefineactions: [
+        {
+          typeAction: 'Refunded fully',
+          amountAction: 100.0,
+          balance: 500.0,
+          accountId: 'r1',
+          userId: 'user1',
+          id: 'w1'
+        },
+        {
+          typeAction: 'Refunded fully',
+          amountAction: 200.0,
+          balance: 400.0,
+          accountId: 'r2',
+          userId: 'user1',
+          id: 'w2'
+        },
+        {
+          typeAction: 'Refunded fully',
+          amountAction: 100.0,
+          balance: 500.0,
+          accountId: 'r3',
+          userId: 'user1',
+          id: 'w3'
+        },
+      ],
+      amount : 400.00
     };
   });
 }