+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Change history for ui-users
 
-## 5.1.0 (IN PROGRESS)
+## 6.0.0 (IN PROGRESS)
 
 * Fix bug showing duplicated service points in add service point dialog. Fixes UIU-1892.
 * Add report icon for report menu items. Refs UIU-1505.
 * Allow search by username. Refs UIU-1707.
 * Use more efficient queries for `accounts` records. Refs UIU-1913.
+* Relocate the username/password in the create/edit and detail view screens. Refs UIU-1035.
+* Allow search by last name. Refs UIU-1706.
+* Add user's fees/fines permissions. Refs UIU-1918.
+* Allow cancel fee/fine as error only if any 'actions' has not been applied. Refs UIU-1894.
+* Allow search by barcode. Refs UIU-1708.
+* Work with multiple and single `refund` action. Refs UIU-1897.
+* Show the number of open requests in `LoanActionDialog`. Refs UIU-1890.
+* Remove optional `action` fields if no value was setted. Refs UIU-1773.
+* Show the number of open requests in the Claim returned bulk action modal. Refs UIU-1891.
 
 ## [5.0.1](https://github.com/folio-org/ui-users/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v5.0.0...v5.0.1)